Switch to pointer implementation
